修复FP16两个潜在BUG：导出顺序+IO Binding类型
- predictor.half()移到ORT导出之后：ONNX导出需要FP32模型+FP32
  dummy input，之前half()在导出前会导致首次运行dtype不匹配崩溃
- ORT IO Binding输入加.float()安全转换：模型FP16后patch可能是
  FP16，但ONNX模型和binding声明FP32，不转换会导致ORT把FP16
  字节当FP32读→输出乱码

+++ src/sharp3d/ort_engine.py
@@ -243,7 +243,9 @@ class ORTEncoder(nn.Module):
     def _forward_iobinding(self, x: torch.Tensor):
         """Zero-copy IO Binding path: GPU tensor → ORT → GPU tensors."""
         batch = x.shape[0]
-        x_contig = x.detach().contiguous()
+        # Ensure FP32: after predictor.half(), patches may arrive as FP16,
+        # but the ONNX model and IO Binding expect FP32 input.
+        x_contig = x.detach().float().contiguous()
 
         # Pre-allocate output tensors on CUDA based on session metadata
         out_tensors = []
